Move VariableChain into JinjavaInterpreter
