+++ src/main/java/com/hubspot/jinjava/lib/filter/SelectAttrFilter.java
@@ -12,7 +12,6 @@ import com.hubspot.jinjava.interpret.JinjavaInterpreter;
 import com.hubspot.jinjava.lib.exptest.ExpTest;
 import com.hubspot.jinjava.util.ForLoop;
 import com.hubspot.jinjava.util.ObjectIterator;
-import com.hubspot.jinjava.util.VariableChain;
 
 @JinjavaDoc(
     value = "Filters a sequence of objects by applying a test to an attribute of an object and only selecting the ones with the test succeeding.",
@@ -56,7 +55,7 @@ public class SelectAttrFilter implements Filter {
     ForLoop loop = ObjectIterator.getLoop(var);
     while (loop.hasNext()) {
       Object val = loop.next();
-      Object attrVal = new VariableChain(Lists.newArrayList(attr), val).resolve();
+      Object attrVal = interpreter.resolve(val, Lists.newArrayList(attr));
 
       if (expTest.evaluate(attrVal, interpreter)) {
         result.add(val);
